Initial commit for the UI
This code controls the UI for "Lauren's Group-Maker"

CURRENTLY:
The user can add/remove students to a master list by selecting
"Add/Remove Student(s)" from a combobox. This list is saved in the
current directory to the text file "studentNames.txt" when the user
clicks the "Save" button.

The user can make groups with the list of students by selecting "Use
This List" from the combobox (this is the default selection). This
changes the text on the button from "Save" to "Make Groups". When the
user clicks "Make Groups", the text file "activityGroups.txt" is saved
to the current directory. This file lists an activity name, followed by
one or more student names. Once all the students for a particular group
have been listed, there is an empty line, followed by the next activity
name, then the students for that activity, etc., until all the
activities are listed with their respective students.

TODO:

generate groups for multiple days

display the group in the UI after they have been generated

OPTIONAL:

allow the user to select the number of days for which they want to
generate groups; e.g. if a week is shortened due to a holiday, the user
should be able to select 4 days

allow the user to add a student directly to a group within the UI. E.g.
if a new student arrives on Wednesday, the user should not have to
remake groups for the entire week, rather they should be able to add the
new student to whatever groups they want, and the student lists (i.e.
the one that is displayed as well as the one in the studentNames.txt
file) should update automatically

allow the user to enter the maximum number of total students (currently
hard-coded at 25)

allow the user to enter the maximum number of students for a single
group (currently hard-coded at 6)

allow the user to add/remove activities