Refactor the class to make two sets of groups
for the first set, half (or half-1) of the names are put into
readToSelf, and the other half are randomly assigned to one of the other
(four) groups. This set is called "firstRotation"
For the second set, all the names from the other (four) groups are put
into readToSelf, and the names originally in readToSelf are randomly
distributed into the other four groups.This set is called
"secondRotation".
Each set is then sent individually to another class for saving to a
file.